0027-use-fw_path-prefix-when-fallback-searching-for-grub-.patch: use fw_path prefix when fallback searching for grub config
When PXE booting via UEFI firmware, grub was searching for grub.cfg
in the fw_path directory where the grub application was found. If
that didn't exist, a fallback search would look for config file names
based on MAC and IP address. However, the search would look in the
prefix directory which may not be the same fw_path. This patch
changes that behavior to use the fw_path directory for the fallback
search. Only if fw_path is NULL will the prefix directory be searched.

grub-2.12 rebase:
  Original patch-name: 0027-use-fw_path-prefix-when-fallback-searching-for-grub-.patch
  Patch modified: no
  Observations: none
  Status: INTERNAL_REVIEW

Signed-off-by: Mark Salter <msalter@redhat.com>