0109-Make-the-menu-entry-users-option-argument-to-be-opti.patch: Make the menu entry users option argument to be optional
The --users option is used to restrict the access to specific menu entries
only to a set of users. But the option requires an argument to either be a
constant or a variable that has been set. So for example the following:

  menuentry "May be run by superusers or users in $users" --users $users {
  	    linux /vmlinuz
  }

Would fail if $users is not defined and grub would discard the menu entry.
Instead, allow the --users option to have an optional argument and ignore
the option if the argument was not set.

Related: rhbz#1652434

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>