efi/net: Allow to specify a port number in addresses
The grub_efi_net_parse_address() function is not covering the case where a
port number is specified in an IPv4 or IPv6 address, so will fail to parse
the network address.

For most cases the issue is harmless, because the function is only used to
match an address with a network interface and if fails the default is used.

But still is a bug that has to be fixed and it causes error messages to be
printed like the following:

error: net/efi/net.c:782:unrecognised network address '192.168.122.1:8080'

error: net/efi/net.c:781:unrecognised network address '[2000:dead:beef:a::1]:8080'

Resolves: rhbz#1732765

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>