net/http: Prepend prefix when the HTTP path is relative as done in efi/http
There are two different HTTP drivers that can be used when requesting an
HTTP resource: the efi/http that uses the EFI_HTTP_PROTOCOL and the http
that uses GRUB's HTTP and TCP/IP implementation.

The efi/http driver appends a prefix that is defined in the variable
http_path, but the http driver doesn't.

So using this driver and attempting to fetch a resource using a relative
path fails.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>