genmod.sh.in: strip .llvm_addrsig sections and similar.
Currently grub modules built with clang or gcc have several sections
which we don't actually need or support.

We already have a list of section to skip in genmod.sh, and this patch
adds the following sections to that list (as well as a few newlines):

.note.gnu.property
.llvm*

Note that the glob there won't work without a new enough linker, but the
failure is just reversion to the status quo, so that's not a big problem.

Signed-off-by: Peter Jones <pjones@redhat.com>