emu/linux: work around systemctl kexec returning
Per systemctl(1), it "is asynchronous; it will return after the reboot
operation is enqueued, without waiting for it to complete".  This
differs from kexec(8), which calls reboot(2) and therefore does not
return.

When not using fallback, this results in the confusing-but-harmless:

    error trying to perform 'systemctl kexec': 0
    Aborted. Press any key to exit.

on screen for a bit, followed by successful kexec.

To reduce the liklihood of hitting this case, add a delay on succesful
return.  Ultimately, the systemd interface is racy: we can't avoid it
entirely unless we never fallback on success.

Signed-off-by: Robbie Harwood <rharwood@redhat.com>