gnulib: Add patch to allow GRUB w/GCC-15 compile
Pull in gnulib fix to allow base64.c to compile using GCC 15 or newer.

Pulled from: GNULIB commit 25df6dc4253480a343dde3376ce6fd99c316a532

GCC 15 adds a new compiler warning "-Wunterminated-string-initialization"
that will trigger what is considered a false-positive in base64.c as this
array is not treated as a string but an array of characters so the lack
of NULL string terminator is expected.

GCC team has added ability to flag such instances of arrays that the
compiler may think are strings as "nonstring" arrays to avoid this
warning: _attribute_((nonstring)).

Fixes: https://savannah.gnu.org/bugs/?66470

Signed-off-by: Andrew Hamilton <adhamilt@gmail.com>